svn: working copy change detection moved to SvnWorkingCopyManager
