+++ de.setsoftware.reviewtool.changesources.svn/src/de/setsoftware/reviewtool/changesources/svn/SvnChangeSource.java
@@ -9,8 +9,6 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.SortedMap;
-import java.util.TreeMap;
 import java.util.regex.Pattern;
 
 import org.eclipse.core.runtime.IProgressMonitor;
@@ -23,10 +21,9 @@ import org.osgi.framework.FrameworkUtil;
 import org.tmatesoft.svn.core.SVNDepth;
 import org.tmatesoft.svn.core.SVNException;
 import org.tmatesoft.svn.core.internal.wc.DefaultSVNAuthenticationManager;
-import org.tmatesoft.svn.core.wc.ISVNStatusHandler;
 import org.tmatesoft.svn.core.wc.SVNClientManager;
 import org.tmatesoft.svn.core.wc.SVNRevision;
-import org.tmatesoft.svn.core.wc.SVNStatus;
+
 import de.setsoftware.reviewtool.base.Pair;
 import de.setsoftware.reviewtool.base.ReviewtoolException;
 import de.setsoftware.reviewtool.model.api.IBinaryChange;
@@ -133,7 +130,7 @@ final class SvnChangeSource implements IChangeSource {
     @Override
     public void analyzeLocalChanges(final List<File> relevantPaths) {
         try {
-            this.collectWorkingCopyChanges(relevantPaths);
+            SvnWorkingCopyManager.getInstance().collectWorkingCopyChanges(relevantPaths);
         } catch (final SVNException e) {
             throw new ReviewtoolException(e);
         }
@@ -211,104 +208,6 @@ final class SvnChangeSource implements IChangeSource {
                 }
             }
         }
-    }
-
-    /**
-     * Collects all local changes and integrates them into the {@link SvnFileHistoryGraph}.
-     * @param relevantPaths The list of additional paths to check. If {@code null}, the whole working copy is analyzed.
-     */
-    private void collectWorkingCopyChanges(final List<File> relevantPaths) throws SVNException {
-        for (final SvnWorkingCopy wc : SvnWorkingCopyManager.getInstance().getWorkingCopies()) {
-            final SortedMap<String, CachedLogEntryPath> changeMap = new TreeMap<>();
-            final ISVNStatusHandler handler = new ISVNStatusHandler() {
-                @Override
-                public void handleStatus(final SVNStatus status) throws SVNException {
-                    if (status.isVersioned()) {
-                        final CachedLogEntryPath entry = new CachedLogEntryPath(wc.getRepository(), status);
-                        changeMap.put(entry.getPath(), entry);
-                    }
-                }
-            };
-
-            if (relevantPaths != null) {
-                final Set<File> filteredPaths = this.filterPaths(relevantPaths, wc);
-                this.collectWorkingCopyChanges(filteredPaths, handler);
-            } else {
-                this.collectWorkingCopyChanges(wc, handler);
-            }
-
-            final SvnWorkingCopyRevision wcRevision = new SvnWorkingCopyRevision(wc, changeMap);
-            final SvnFileHistoryGraph localFileHistoryGraph = new SvnFileHistoryGraph();
-            localFileHistoryGraph.processRevision(wcRevision);
-            wc.setLocalFileHistoryGraph(localFileHistoryGraph);
-        }
-    }
-
-    /**
-     * Collects local changes given a set of paths.
-     * @param paths The paths to consider.
-     * @param handler Receives information about changes files.
-     */
-    private void collectWorkingCopyChanges(final Set<File> paths, final ISVNStatusHandler handler)
-            throws SVNException {
-
-        for (final File path : paths) {
-            this.mgr.getStatusClient().doStatus(
-                    path,
-                    SVNRevision.WORKING,
-                    SVNDepth.EMPTY,
-                    false, // no remote
-                    false, // report only modified files
-                    false, // don't include ignored files
-                    false, // ignored
-                    handler,
-                    null); // no change lists
-        }
-    }
-
-    /**
-     * Collects local changes within a whole working copy.
-     * @param wc The working copy to consider.
-     * @param handler Receives information about changes files.
-     */
-    private void collectWorkingCopyChanges(final SvnWorkingCopy wc, final ISVNStatusHandler handler)
-            throws SVNException {
-
-        this.mgr.getStatusClient().doStatus(
-                wc.getLocalRoot(), // analyse whole working copy
-                SVNRevision.WORKING,
-                SVNDepth.INFINITY,
-                false, // no remote
-                false, // report only modified files
-                false, // don't include ignored files
-                false, // ignored
-                handler,
-                null); // no change lists
-    }
-
-    /**
-     * Filters out paths that do not belong to passed working copy.
-     * @param relevantPaths The paths to filter.
-     * @param wc The relevant working copy.
-     * @return A set of filtered paths.
-     */
-    private Set<File> filterPaths(final List<File> relevantPaths, final SvnWorkingCopy wc) {
-        final Set<File> paths = new LinkedHashSet<>();
-        for (final File path : relevantPaths) {
-            final String repoPath = wc.toAbsolutePathInRepo(path);
-            if (repoPath != null) {
-                paths.add(path);
-            }
-        }
-
-        for (final String repoPath : wc.getLocalFileHistoryGraph().getPaths()) {
-            final File path = wc.toAbsolutePathInWc(repoPath);
-            if (path != null && path.isFile()) {
-                paths.add(path);
-            }
-        }
-
-        return paths;
     }
 
     private Map<ISvnRepo, Long> determineMaxRevisionPerRepo(
